+++ lib/p2/compiler.rb
@@ -4,203 +4,10 @@ require 'cgi'
 require 'sirop'
 require 'erb/escape'
 
+require_relative './compiler/nodes'
+require_relative './compiler/tag_translator'
+
 module P2
-  class TagNode
-    attr_reader :call_node, :location, :tag, :tag_location, :inner_text, :attributes, :block
-
-    def initialize(call_node, transformer)
-      @call_node = call_node
-      @location = call_node.location
-      @tag = call_node.name
-      prepare_block(transformer)
-
-      args = call_node.arguments&.arguments
-      return if !args
-
-      if @tag == :tag
-        @tag = args[0]
-        args = args[1..]
-      end
-
-      if args.size == 1 && args.first.is_a?(Prism::KeywordHashNode)
-        @inner_text = nil
-        @attributes = args.first
-      else
-        @inner_text = args.first
-        @attributes = args[1].is_a?(Prism::KeywordHashNode) ? args[1] : nil
-      end
-    end
-
-    def accept(visitor)
-      visitor.visit_tag_node(self)
-    end
-
-    def prepare_block(transformer)
-      @block = call_node.block
-      if @block.is_a?(Prism::BlockNode)
-        @block = transformer.visit(@block)
-        offset = @location.start_offset
-        length = @block.opening_loc.start_offset - offset
-        @tag_location = @location.copy(start_offset: offset, length: length)
-      else
-        @tag_location = @location
-      end
-    end
-  end
-
-  class RenderNode
-    attr_reader :call_node, :location, :block
-
-    include Prism::DSL
-
-    def initialize(call_node, transformer)
-      @call_node = call_node
-      @location = call_node.location
-      @transformer = transformer
-      @block = call_node.block && transformer.visit(call_node.block)
-
-      lambda = call_node.arguments && call_node.arguments.arguments[0]
-      return unless lambda.is_a?(Prism::LambdaNode)
-
-      location = lambda.location
-      parameters = lambda.parameters
-      parameters_location = parameters&.location || location
-      params = parameters&.parameters
-      lambda = lambda_node(
-        location: location,
-        parameters: block_parameters_node(
-          location: parameters_location,
-          parameters: parameters_node(
-            location: parameters_location,
-            requireds: [
-              required_parameter_node(
-                location: ad_hoc_string_location('__buffer__'),
-                name: :__buffer__
-              ),
-              *params&.requireds
-            ],
-            optionals: transform_array(params&.optionals),
-            rest: transform(params&.rest),
-            posts: transform_array(params&.posts),
-            keywords: transform_array(params&.keywords),
-            keyword_rest: transform(params&.keyword_rest),
-            block: transform(params&.block)
-          )
-        ),
-        body: transformer.visit(lambda.body)
-      )
-      call_node.arguments.arguments[0] = lambda
-      # pp lambda_body: call_node.arguments.arguments[0]
-    end
-
-    def ad_hoc_string_location(str)
-      src = source(str)
-      Prism::DSL.location(source: src, start_offset: 0, length: str.bytesize)
-    end
-
-    def transform(node)
-      node && @transformer.visit(node)
-    end
-
-    def transform_array(array)
-      array ? array.map { @transformer.visit(it) } : []
-    end
-
-    def accept(visitor)
-      visitor.visit_render_node(self)
-    end
-  end
-
-  class TextNode
-    attr_reader :call_node, :location
-
-    def initialize(call_node, _compiler)
-      @call_node = call_node
-      @location = call_node.location
-    end
-
-    def accept(visitor)
-      visitor.visit_text_node(self)
-    end
-  end
-
-  class RawNode
-    attr_reader :call_node, :location
-
-    def initialize(call_node, _compiler)
-      @call_node = call_node
-      @location = call_node.location
-    end
-
-    def accept(visitor)
-      visitor.visit_raw_node(self)
-    end
-  end
-
-  class DeferNode
-    attr_reader :call_node, :location, :block
-
-    def initialize(call_node, compiler)
-      @call_node = call_node
-      @location = call_node.location
-      @block = call_node.block && compiler.visit(call_node.block)
-    end
-
-    def accept(visitor)
-      visitor.visit_defer_node(self)
-    end
-  end
-
-  class CustomTagNode
-    attr_reader :tag, :call_node, :location, :block
-
-    def initialize(call_node, compiler)
-      @call_node = call_node
-      @tag = call_node.name
-      @location = call_node.location
-      @block = call_node.block && compiler.visit(call_node.block)
-    end
-
-    def accept(visitor)
-      visitor.visit_custom_tag_node(self)
-    end
-  end
-
-  class TagTransformer < Prism::MutationCompiler
-    include Prism::DSL
-
-    def self.transform(ast)
-      ast.accept(new)
-    end
-
-    def visit_call_node(node)
-      # We're only interested in compiling method calls without a receiver
-      return super(node) if node.receiver
-
-      case node.name
-      when :emit_yield
-        yield_node(
-          location: node.location,
-          arguments: node.arguments
-        )
-      when :raise
-        super(node)
-      when :render
-        RenderNode.new(node, self)
-      when :raw
-        RawNode.new(node, self)
-      when :text
-        TextNode.new(node, self)
-      when :defer
-        DeferNode.new(node, self)
-      when :html5, :markdown
-        CustomTagNode.new(node, self)
-      else
-        TagNode.new(node, self)
-      end
-    end
-  end
-
   class VerbatimSourcifier < Sirop::Sourcifier
     def visit_tag_node(node)
       visit(node.call_node)
@@ -215,7 +22,7 @@ module P2
       ast = ast.block if ast.is_a?(Prism::CallNode)
 
       compiler = new.with_source_map(proc, ast)
-      transformed_ast = TagTransformer.transform(ast.body)
+      transformed_ast = TagTranslator.transform(ast.body)
       compiler.format_compiled_template(transformed_ast, ast, wrap:, binding: proc.binding)
       [compiler.source_map, compiler.buffer]
     end
@@ -409,7 +216,7 @@ module P2
       emit("}")
     end
 
-    def visit_custom_tag_node(node)
+    def visit_builtin_node(node)
       case node.tag
       when :tag
         args = node.call_node.arguments&.arguments
